Дополнил метод getEvent теперь он заполняет все поля сущности event

+++ ewm-service/src/main/java/ru/practicum/users/repository/ParticipationRequestRepository.java
@@ -19,5 +19,5 @@ public interface ParticipationRequestRepository extends JpaRepository<Participat
     Optional<ParticipationRequest> findByIdAndUserId(Long id, Long userId);
 
     @Query("SELECT COUNT(pr) FROM ParticipationRequest pr WHERE pr.status = :status AND pr.event.id = :eventId")
-    long countConfirmedRequestsByStatusAndEventId(@Param("status") ParticipationRequestStatus status, @Param("eventId") Long eventId);
+    int countConfirmedRequestsByStatusAndEventId(@Param("status") ParticipationRequestStatus status, @Param("eventId") Long eventId);
 }
